+++ dk_essentials.h
@@ -12,7 +12,7 @@ void Gamespy_Create_Challenge_Key(char *s, const int len);
 const char *Gamespy_Get_Game_SecKey (char *gamename);
 unsigned short Gamespy_Get_MOTD_Port (char *gamename);
 char *DK_strlwr (char *s); // FS: Some compilers might not have this
-void Parse_ServerList (unsigned int fileSize, char *fileBuffer, char *gamenameFromHttp);
+void Parse_ServerList (size_t fileSize, char *fileBuffer, char *gamenameFromHttp);
 void AddServers_From_List_Execute(char *fileBuffer, char *gamenameFromHttp); // FS: From Quake 2
 
 /* FS: Aluigi.org stuff */
